refactor: SwiftLint 違反を解消する（file_length / function_body_length / cyclomatic_complexity）
- ChatSendError を ChatViewModel.swift から ChatSendError.swift に分離
  → ChatViewModel.swift の file_length を 807→687 行に削減
- startStreamTasks の USERSTATE 処理を handleUserStateUpdate(_:) に抽出
  → function_body_length を 51→38 行、cyclomatic_complexity を 17→13 に削減
- fetchUserEmotes のページネーションループを performUserEmotesFetch(userId:) に抽出
  → function_body_length を 53→36 行に削減
